+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ next
 - Fixed support for Matplotlib 3.5.
 - Bumped supported raqm versions to ≥0.7.
 - Support Qt6 and GTK4.
+- Support HiDPI on GTK (it was already supported on Qt).
 
 v0.4
 ====